add changes from feedback
