+++ db.js
@@ -1,0 +1,543 @@
+const asanas = [
+  {
+    id: 1,
+    sanskrit_name: "Navasana",
+    english_name: "Boat",
+    img_url: "/images/boat.svg",
+    description: `Die Übung zählt zu den wichtigsten klassischen Übungen im Hatha Yoga. Der Begriff Navasana stammt aus dem Sanskrit und ist eine Verbindung der beiden Wörter Nava (Boot) und Asana (Körperstellung).
+      Die Übung hat zwei zentrale Wirkungen. Sie dient einerseits der Stärkung der Bauchmuskeln und andererseits einer Aktivierung des Solarplexus.
+      Die Boothaltung gehört in den Bereich der Gleichgewichtsübungen und ermöglicht Dir eine Stabilisierung Deines unteren Rückens und eine Vorbeugung eines Hohlkreuzes.`,
+    level: ["intermediate", "pro"],
+    category: "Core",
+    benefit: "energizing",
+    next: ["Downward-Facing Dog", "Half-Boat"],
+  },
+  {
+    id: 2,
+    sanskrit_name: "Ardha Navasana",
+    english_name: "Half-Boat",
+    img_url: "/images/half-boat.svg",
+    description: `Navasana wird wegen der Form, die der Körper dabei annimmt, auch Halbboot-Pose genannt. Es wird beschrieben als Ardha bedeutet halb, Nava bedeutet Boot und Asana bedeutet Körperhaltung. Es enthält Variationen wie Paripurna Navasana - Vollständige Bootshaltung, Ardha Navasana - Halbe Bootshaltung und Ekapada Navasana - Einbeinige Bootshaltung.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Core",
+    benefit: "energizing",
+    next: ["Bridge"],
+  },
+  {
+    id: 3,
+    sanskrit_name: "Dhanurasana",
+    english_name: "Bow",
+    img_url: "/images/bow.svg",
+    description: `In dieser Asana formen Sie mit Ihrem Körper einen weiten Bogen. Wenn Sie diese Asana häufig praktizieren, bleibt Ihr Körper beweglich und erleben ein Gefühl von Leichtigkeit. DIe Übung stärkt die WIrbelsäule sowie die Organe des Bauches.`,
+    level: ["intermediate", "pro"],
+    category: "Rücken",
+    benefit: "energizing",
+    next: ["Camel"],
+  },
+  {
+    id: 4,
+    sanskrit_name: "Setu Bandha Sarvangasana",
+    english_name: "Bridge",
+    img_url: "/images/bridge.svg",
+    description: `Die Yoga Schulterbrücke weckt das Bewusstsein und den Geist, und verschärft die Sinne und den Blick: Das sind nur einige Effekte, sterben diese Hatha Yoga Übung Setu Bandha Sarvangasana auf dich hat. Die Yoga Schulterbrücke ist gut für dein Herz und deine Lunge, sie gibt dir neue Energie und beruhigt gleichzeitig deine Nerven und deinen Geist. Die Yoga Schulterbrücke ist eine Mischung aus Halasana (der Pflug) und Salamba Sarvangasana (der Schulterstand). Diese Asanas können super in Kombination geübt werden.`,
+    level: ["intermediate", "pro"],
+    category: "Rücken",
+    benefit: "calming",
+    next: ["Shoulder Stand", "Upward-Facing Dog", "Plow"],
+  },
+  {
+    id: 5,
+    sanskrit_name: "Baddha Konasana",
+    english_name: "Butterfly",
+    img_url: "/images/butterfly.svg",
+    description: `Die Butterfly Pose ist ein klassischer Hüftöffner und wird auch Bound Angle Pose oder Cobbler Pose genannt. Auf Sanskrit ist es als Baddha Konasana bekannt. Die Schmetterlingshaltung ist für alle Niveaus geeignet und daher eine nützliche Ergänzung zu den meisten Yoga-Routinen. Da die Haltung die Flexibilität verbessert und Verspannungen reduziert, ist sie ideal für Menschen mit verspannten Hüften aufgrund von übermäßigem Sitzen oder hochintensivem Training.
+      Es fördert auch das Gefühl der Ruhe und fördert die innere Achtsamkeit.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Hüften",
+    benefit: "calming",
+    next: ["Seated Forward Bend", "Forward Bend with Shoulder Opener"],
+  },
+  {
+    id: 6,
+    sanskrit_name: "Ustrasana",
+    english_name: "Camel",
+    img_url: "/images/camel.svg",
+    description: `In dieser Übung beugt man den Rücken soweit nach hinten, bis der Körper die Form eines Kamels annimt. Diese Asana eigenet sich besonders gut für Menschen, die im Sitzen arbeiten und sich dabei häufig nach vorne beugen. Die Kamelhaltung stärkt die Muskulatur des Rückens und löst Verspannungen in Schultern, Rücken und Fußgelenken. Das Wichtigste, was Sie bei Ustrasana vermeiden sollten, ist, in die Pose zu fallen und die Hauptlast der Rückbeuge in Ihrem Nacken oder unteren Rücken zu tragen. Heben und verlängern Sie stattdessen Ihren Oberkörper, bevor Sie sich sanft in die Pose biegen.`,
+    level: ["intermediate", "pro"],
+    category: "Rücken",
+    benefit: "energizing",
+    next: ["Child's Pose", "Pigeon"],
+  },
+  {
+    id: 7,
+    sanskrit_name: "Marjaryasana",
+    english_name: "Cat",
+    img_url: "/images/cat.svg",
+    description: `Marjaryasana (Katzenhaltung) und seine Gegenhaltung, die Kuhhaltung (Bitilasana), begegnen Ihnen in vielen Yogastunden. Und das aus gutem Grund: Diese Posen helfen, die Wirbelsäule aufzuwärmen. Die Katzenhaltung kann als Vorbereitung für Vorwärtsbeugen und als expansive Gegenhaltung zu Inversionen und Rückbeugen dienen.
+      Konzentrieren Sie sich in dieser Pose darauf, Ihr Steißbein einzustecken, Ihre Wirbelsäule zu runden und Ihren Nacken zu entspannen. Wenn Sie sich dieser Pose vollständig hingeben und durch sie hindurch atmen wird sich Ihr Körper wohl fühlen.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Rücken",
+    benefit: "calming",
+    next: ["Cow"],
+  },
+  {
+    id: 8,
+    sanskrit_name: "Bitilasana",
+    english_name: "Cow",
+    img_url: "/images/cow.svg",
+    description: `Bitilasana (Kuhhaltung), eine Rückbeuge, wird üblicherweise zu Beginn eines Vinyasa-Flusses mit Marjaryasana (Katzenhaltung) kombiniert, um Ihren Körper insbesondere Ihre Wirbelsäule für andere Posen aufzuwärmen. Cow hilft, Verspannungen im Oberkörper zu lösen, insbesondere in Rücken, Schultern und Nacken, und massiert sanft die Wirbelsäule, um die Beweglichkeit zu erhöhen.
+      Diese einfache Pose ermöglicht eine starke Dehnung des Vorderkörpers vom Kinn bis zum Schambein. Heben Sie Ihr Steißbein, um Ihren Rücken zu wölben, lassen Sie Ihren Bauch zum Boden hängen, ziehen Sie Ihre Schultern von Ihren Ohren weg und heben Sie Ihren Kopf.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Rücken",
+    benefit: "calming",
+    next: ["Cat", "Sphinx"],
+  },
+  {
+    id: 9,
+    sanskrit_name: "Utkatasana",
+    english_name: "Chair",
+    img_url: "/images/chair.svg",
+    description: `Utkatasana (Stuhlhaltung) wird manchmal aus dem Sanskrit ins Englische als „heftiger Sitz“ oder „kraftvolle Haltung“ übersetzt. Es ist eine stärkende und wärmeaufbauende Asana, die alle Teile deines Körpers zu einem zusammenhängenden und kraftvollen Ganzen zusammenführt.
+      Dieses Asana ist eine Meditation über Entschlossenheit und Ausdauer sowie Hingabe. Um die Stuhlhaltung erfolgreich auszuführen, müssen Sie die Kraft der Beine, Arme und des Oberkörpers nahtlos vereinen, während Sie Ihre Rumpfmuskulatur anheben und Ihre Wirbelsäule verlängern.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Beine",
+    benefit: "energizing",
+    next: ["Standing Forward Bend", "Bridge", "Downward-Facing Dog"],
+  },
+  {
+    id: 10,
+    sanskrit_name: "Balasana",
+    english_name: "Child's Pose",
+    img_url: "/images/child.svg",
+    description: `Die Kinderhaltung (Balasana) ist die wichtigste Ruhehaltung im Yoga und eine schöne Möglichkeit, verschiedene Körperteile sanft zu dehnen.
+      Es ist eine Gelegenheit, mit dem, was Sie tun, aufzuhören, Ihre Position neu zu bewerten, sich wieder mit Ihrem Atem zu verbinden und sich darauf vorzubereiten, vorwärts zu gehen.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Rücken",
+    benefit: "calming",
+    next: ["Dolphin", "Downward-Facing Dog"],
+  },
+  /*  {
+      id: 11,
+      sanskrit_name: "Sivasana",
+      english_name: "Corpse",
+      img_url: "",
+      description: `Die Leiste ist entspannt, die Wirbelsäule lang und die Arme liegen im 30 bis 45° Winkel neben dem Oberkörper. Die Handflächen zeigen nach oben, sodass sich die Brust und Schultern öffnen. Die Schultern dürfen nicht hochgezogen sein. Ein wichtiges Merkmal ist die Wachsamkeit. Du liegst auf dem Rücken völlig still und bewegungslos und hältst die Aufmerksamkeit vollkommen bei Shavasana und nimmst den Augenblick gänzlich wahr. Der Körper wird ganz leise, nur der Moment ist präsent. Er liegt starr auf dem Boden, sodass sich nicht mal ein Finger bewegt`,
+      level: ["beginner", "intermediate", "pro"],
+      category: "Rücken",
+      benefit: "calming",
+      next: [],
+    }, */
+  {
+    id: 12,
+    sanskrit_name: "Alanasana",
+    english_name: "Crescent Lunge",
+    img_url: "/images/crescent_lunge.svg",
+    description: `Alanasana, der hohe Ausfallschritt ist in vielen verschiedenen Sportarten bekannt. Im Yoga wird er häufig mit dem Krieger I verwechselt. Der Unterschied besteht vor allem in der Ausrichtung des hinteren Beins und Fußes. Die Ferse ist angehoben, was auch Einfluss auf die Haltung der Hüfte hat. Der Oberkörper wird weiter nach hinten gebeugt, als in der Kriegerhaltung. So werden im Crescent Lunge die Hüftflexoren stärker gedehnt. Diese Haltung ist insbesondere für Läufer gut geeignet. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Hüfte",
+    benefit: "calming",
+    next: ["Plank", "Downward-Facing Dog", "Standing Forward Bend"],
+  },
+  {
+    id: 13,
+    sanskrit_name: "Bakasana",
+    english_name: "Crow",
+    img_url: "/images/crow.svg",
+    description: `Die Krähen-Haltung im Yoga steht für starke Arme und perfekte Balance. In dieser Asana balancierst du das Gewicht deines Körpers auf deinen Händen. Die Arme sind dabei leicht gebeugt, die Knie liegen auf dem Trizeps auf und die Füße schweben in der Luft. Das Gesäß ist der höchste Punkt deines Körpers bei dieser Übung. Die Haltung erfordert vor allem einen starken Core und viel Balance.`,
+    level: ["intermediate", "pro"],
+    category: "Schultern",
+    benefit: "calming",
+    next: ["Squat", "Butterfly", "Child's Pose"],
+  },
+  {
+    id: 14,
+    sanskrit_name: "Ardha Pincha Mayurasana",
+    english_name: "Dolphin",
+    img_url: "/images/dolphin.svg",
+    description: `Die Delfin-Pose ist eine Haltung zur Stärkung der Arme und der Schultern. Sie dient in viele Fällen der Vorbereitung auf andere, eher fortgeschrittene Posen. Diese Übung ist eine Variation vom herabschauenden Hund, bei der die Unterarme auf dem Boden aufliegen. Die Füße stehen fest auf dem Boden, während der Steiß nach oben gehoben wird. Die Delfin-Haltung öffnet die Schultern, verlängert die Wirbelsäule und streckt die Beinrückseiten.`,
+    level: ["intermediate", "pro"],
+    category: "Schultern",
+    benefit: "energizing",
+    next: ["Downward-Facing Dog", "Plank", "Forearm Stand"],
+  },
+  {
+    id: 15,
+    sanskrit_name: "Adho Mukha Svanasana",
+    english_name: "Downward-Facing Dog",
+    img_url: "/images/downdog.svg",
+    description: `In dieser Asana nimmt der Körper die Form eines sich an unten streckenden Hundes an. Diese Übung dehnt die Beinrückseiten und stärkt gleichzeitig die Beinmuskulatur. Auch die Schulter und Arme werden gekräftigt. Der herabschauende Hund bringt frische Energie und stimuliert das Nervensytem. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Schultern",
+    benefit: "calming",
+    next: ["Plank", "Low Lunge", "Upward-Facing Dog"],
+  },
+  {
+    id: 16,
+    sanskrit_name: "Garudasana",
+    english_name: "Eagle",
+    img_url: "/images/eagle.svg",
+    description: `Die Adler-Haltung im Yoga ist eine Übung für mehr Konzentration und Gleichgewicht. Bei dieser stehenden Balance-Haltung werden sowohl die Beine als auch die Arme ineinander verschlungen und der Körper leicht nach vorne geneigt. Der Adler dehnt den Bereich zwischen den Schultern und stärkt gleichzeitig die Muskulatur der Oberschenkel. Verspannung in Hüfte und Kreuzbein werden gelöst.`,
+    level: ["intermediate", "pro"],
+    category: "Schultern",
+    benefit: "calming",
+    next: ["Chair", "Tree"],
+  },
+  {
+    id: 17,
+    sanskrit_name: "Utthita Hasta Padangusthasana",
+    english_name: "Extended Hand to Toe",
+    img_url: "/images/extendedhandtotoe.svg",
+    description: `Diese Pose, auf deutsch auch erweiterte Hand-zu-Großzeh-Pose genannt, ist eine ausgleichende Haltung, bei der Sie ein Bein auf Hüfthöhe anheben und mit Ihrer Hand an seinem großen Zeh festhalten. Mit dieser Asana werden Kraft und Flexibilität aufgebaut, insbesondere in den Beinen und Knöcheln. Sie stärkt den Fokus, das Selbstvertrauen und die Balance.`,
+    level: ["intermediate", "pro"],
+    category: "Hüfte",
+    benefit: "calming",
+    next: ["Downward-Facing Dog", "Tree"],
+  },
+  {
+    id: 18,
+    sanskrit_name: "Utthita Parsvakonasana",
+    english_name: "Extended Side Angle",
+    img_url: "/images/extendedsideangle.svg",
+    description: `In dieser Asana werden beide Seiten des Körpers intensiv gedehnt. Der Körper wird von den Zehen wir zu den Fingerspitzen durchgestreckt. Dabei ist es wichtig, den Körper in gleichmäßiger Ausrichtung zu halten und nicht zu sehr in das gebeugte Bein sinken zu lassen. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Hüfte",
+    benefit: "energizing",
+    next: [""],
+  },
+  {
+    id: 19,
+    sanskrit_name: "Pincha Mayurasana",
+    english_name: "Forearm Stand",
+    img_url: "/images/forearmstand.svg",
+    description: `Der Unterarmstand ist eine Umkehrhaltung für Fortgeschrittene. Sie stärkt Arme, Schultern, Rumpf und Rücken. Als Umkehrung erhalten Sie eine erhöhte Durchblutung des Gehirns, was helfen kann, Stress abzubauen. Die Übung trainiert das Gleichgewicht und fördert einen ruhigen Geist. Zudem können sich Inversionen, wie der Unterarmstand positiv auf das kardiovaskuläre, lymphatische, nervöse und endokrine System auswirken.`,
+    level: ["pro"],
+    category: "Schulter",
+    benefit: "calming",
+    next: ["Handstand", "Child's Pose"],
+  },
+  {
+    id: 20,
+    sanskrit_name: "Uttanasana",
+    english_name: "Forward Bend with Shoulder Opener",
+    img_url: "/images/forwardfoldshoulderstretch.svg",
+    description: `Die stehende Vorwärtsbeuge mit Schulteröffnung ist eine Variation der klassichen Vorwärtsbeuge. Dabei werden die Hände hinter dem Rücken verschränkt und beim Vorbeugen über den Kopf fallen lassen. Die öffnet die Schultern und dehnt den Bereich zwischen den Schulterblättern. Zudem werden die Beinrückseiten gedehnt und die gesamte Wirbelsäule gestreckt.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Schulter",
+    benefit: "calming",
+    next: ["Cat", "Cow", "Downward-Facing Dog"],
+  },
+  {
+    id: 21,
+    sanskrit_name: "Ardha Chandrasana",
+    english_name: "Half-Moon",
+    img_url: "/images/halfmoon.svg",
+    description: `Der Halb-Mond ist eine herausfordernde Übung, die aber auch Anfänger praktizieren können. Die Haltung symbolisiert, wie der Name bereist verrät, einen halben Mond. Sie erfordert Balance, einen sicheren Stand und viel Kraft aus der Körpermitte. In dieser Asana werden die Beine gestreckt, die Hüfte gedehnt und der Brustkorb geweitet. Sie ist besonders wirkungsvoll gegen Schmerzen im unteren Rücken. Regelmäßiges Üben bringt zudem den Kreislauf in Schwung und verbessert die Atmung. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Schultern",
+    benefit: "calming",
+    next: [""],
+  },
+  {
+    id: 22,
+    sanskrit_name: "Adho Mukha Vrksasana",
+    english_name: "Handstand",
+    img_url: "/images/handstand.svg",
+    description: `Der Handstand ist eine der herausfordernsten Übungen im Yoga. Die Asana stärkt Arme, Schulter und Rücken und dehnt Bauch, Hüften und Oberschenkel. Auch die Körpermitte, der Core ist bei dieser Haltung stark gefordert. Die Haltung erfordert Mut und Vertrauen in das eigene Können. Sie ist nicht nur eine körperliche, sondern auch eine mentale Herausforderung.`,
+    level: ["pro"],
+    category: "Schultern",
+    benefit: "energizing",
+    next: ["Child's Pose", "Bridge", "Bow"],
+  },
+  {
+    id: 23,
+    sanskrit_name: "Anjaneyasana",
+    english_name: "Low Lunge",
+    img_url: "/images/lowlunge.svg",
+    description: `Der Low Lunge oder auch tiefer Ausfallschritt genannt, ist eine Haltung die Oberschenkel, Leiste und Hüfte dehnt und gleichzeitig den Brustkorb öffnet. Um dem Rücken nicht zu schaden, soll bewusst darauf geachtet werden, nicht ins Hohlkreuz zu verfallen. Das Knie sollte sich idealerweise direkt über dem Sprunggelenk befinden.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Hüfte",
+    benefit: "energizing",
+    next: ["Plank", "Warrior One", "Warrior Two"],
+  },
+
+  {
+    id: 24,
+    sanskrit_name: "Supta Kapotasana",
+    english_name: "Pigeon",
+    img_url: "/images/pigeon.svg",
+    description: `Die Taube ist eine intensive Hüftöffung im Sitzen. Gerade nach einem langen Tag am Schreibtisch wirkt diese Pose ausgleichend, beruhigend und wohltuend. Die Taube sorgt durch die Außenrotation und die Beugung der Hüfte dafür, dass das Gelenk beweglich und flexibel bleibt. Verspannungen werden gelöst und der Psoas-Muskel im Gesäß wird gedehnt. So können auch Schmerzen im unteren Rücken vorgebeugt werden.`,
+    level: ["intermediate", "pro"],
+    category: "Hüfte",
+    benefit: "calming",
+    next: ["Camel", "Butterfly", "King Pigeon"],
+  },
+  {
+    id: 25,
+    sanskrit_name: "Eka Pada Rajakapotasana",
+    english_name: "King Pigeon",
+    img_url: "/images/kingpigeon.svg",
+    description: `Die einbeinige Königstaube ist eine Variation der Taube. Sie ist eine intensive Hüftöffnung kombiniert mit einer tiefen Rückbeuge. Die Hüfte als zentrales Bewegungszentrum unseres Körpers wird durch das viele Sitzen häufig eng und unflexibel. Durch diese Asana bringen wird die Durchblutung in der Hüfte, den Leisten, der Oberschenkelmuskulatur verstärkt und so die Beweglichkeit gefördert. Die Wirbelsäule wir durch diese Haltung gestreckt und entlastet.`,
+    level: ["pro"],
+    category: "Hüfte",
+    benefit: "calming",
+    next: ["Squad", "Butterfly", "Pigeon"],
+  },
+  {
+    id: 26,
+    sanskrit_name: "Phalakasana",
+    english_name: "Plank",
+    img_url: "/images/plank.svg",
+    description: `Die Plank-Haltung, auch bekannt als Yoga-Liegestütz ist eine Übung die den ganzen Körper fordert. Sie ein ausgezeichnetes Bauch- und Rumpftraining. Sie stärkt die Wirbelsäule, sowie Arme und Schultern. Achten Sie darauf, dass Ihre Hüften nicht zu weit absinken. Der Kopf und Nacken solte sich mit dem Körper in einer Linie befinden.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Schultern",
+    benefit: "energizing",
+    next: ["Upward-Facing Dog", "Sphinx", "Child's Pose"],
+  },
+  {
+    id: 27,
+    sanskrit_name: "Halasana",
+    english_name: "Plow",
+    img_url: "/images/plow.svg",
+    description: `In dieser Asana formt der Körper eine Pflug. Regelmäßiges Üben von Halasana stärkt das Selbstbewusstsein und verleiht Enegerie. Da Halasana die Augen und das Gehirn entspannt, lindert es die Auswirkungen von Stress und psychischen Belastungen, Die Überung streckt die Wirbelsäule und fördert die Verdauuung. `,
+    level: ["intermediate", "pro"],
+    category: "Schultern",
+    benefit: "energizing",
+    next: ["Shoulder Stand", "Corpse"],
+  },
+  {
+    id: 28,
+    sanskrit_name: "Parsvottanasana",
+    english_name: "Pyramid",
+    img_url: "/images/pyramid.svg",
+    description: `Die Pyramide zählt zu den stehenden Yoga-Asanas. Sie symbolisiert eine stabile und starke Basis und strahlt Energie und Kraft aus. Stabilität und Gleichgewicht werden dabei ebenso gestärkt wie die Muskulatur der Beine und des Oberkörpers. Auch Becken, Hüfte, Rücken und Bauch werden bei der Pyramide trainiert. Die Wirbelsäule wie auch die Flanken werden gedehnt und gestreckt.`,
+    level: ["intermediate", "pro"],
+    category: "Core",
+    benefit: "energizing",
+    next: ["Seated Forward Bend", "Standing Forward Bend"],
+  },
+  {
+    id: 29,
+    sanskrit_name: "Parsva Virabhadrasana",
+    english_name: "Reverse Warrior",
+    img_url: "/images/reverswarrior.svg",
+    description: `Der umgedrehte Krieger, oder auch friedlicher Krieger genannt, ist eine Variation des Kriegers II. Dabei geht der hintere Arm an das hintere Bein, der Oberkörper beugt sich nach hinten und der vorder Arm zeigt nach oben. Der Blick folgt der vorderen Hand. Der Reverse Warrior stärkt Arme und Beine und verbessert die Flexibilität der Wirbelsäule. Ebenso wird die Durchblutung im gesamten Körper verbessert. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Beine",
+    benefit: "energizing",
+    next: ["Warrior Two", "Pyramid"],
+  },
+
+  {
+    id: 30,
+    sanskrit_name: "Paschimottanasana",
+    english_name: "Seated Forward Bend",
+    img_url: "/images/seatedforwardfold.svg",
+    description: `Diese Haltung gehört zu den sitzenden Asanas. Die Beine sind werden lang nach vorne ausgestreckt, die Füße sind eng zusammen. Die Arme werden erst bei der Einatmung nach oben gestreckt, anschließend wird die Hüfte gebeugt und der Oberkörper legt sich auf den Oberschenkeln ab. Nacken und Wirbelsäule bleiben dabei in einer Linie. Die Oberschenkelrückseiten werden bei dieser Übung intensiv gedehnt. Ebenso werden Wirbelsäule, Hüftbeuger und unterere Rücken gestreckt. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Beine",
+    benefit: "calming",
+    next: ["Seated Spinal Twist", "Butterfly"],
+  },
+  {
+    id: 31,
+    sanskrit_name: "Padmasana",
+    english_name: "Lotus",
+    img_url: "/images/lotus.svg",
+    description: `Der Lotus-Sitz gehört zu den bekanntesten Yoga-Asanas. Er gilt als optimale Sitzposition für Meditationen. Die Wirbelsäule ist in dieser Haltung aufgerichtet, so treten auch bei längerem Sitzen keine Rückenbeschwerden auf. Auch die Hüftmuskulatur wird gedehnt. Die Haltung beruhigt Körper und Geist. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Beine",
+    benefit: "calming",
+    next: ["Seated Spinal Twist", "Butterfly"],
+  },
+  {
+    id: 32,
+    sanskrit_name: "Ardha Matsyendrasana",
+    english_name: "Seated Spinal Twist",
+    img_url: "/images/seatedspinaltwist.svg",
+    description: `Diese Übung ist auch unter dem Namen Rumpfdrehung bekannt. Sie eine tolle Übung für Anfänger, die gezielt Ihre Rückenmuskulatur stärken möchten sowie die Beweglichkeit ihrer Wirbelsäule verbessern möchten. Der mittlere Rücken wird gedehnt und die inneren Bauchorgane leicht massiert. Die Übung öffnet zudem den Brust-, Herz und Lungenraum. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Rücken",
+    benefit: "calming",
+    next: ["Seated Forward Bend", "Butterfly"],
+  },
+  {
+    id: 33,
+    sanskrit_name: "Salamba Sarvangasana",
+    english_name: "Shoulder Stand",
+    img_url: "/images/shoulderstand.svg",
+    description: `Der Schulterstand gehört zu den wichtigesten Yoga-Haltungen und wird auch als Königin der Asanas bezeichnet. Die Übung zählt zu den Umkehrhaltungen. Das gesamte Körpergewicht wird von den Schultern und Armen getragen. Die Beine und das Becken befinden sich in der Luft in einer gerade Linie über dem Herzen. Diese Übung eignet sich besonders für dem Schlafen-Gehen oder zur Regeneration. `,
+    level: ["intermediate", "pro"],
+    category: "Schulter",
+    benefit: "calming",
+    next: ["Corpse", "Plow"],
+  },
+
+  {
+    id: 34,
+    sanskrit_name: "Vasisthasana",
+    english_name: "Side Plank",
+    img_url: "/images/sideplank.svg",
+    description: `Vasisthasana, oder Side Plank Pose, ist eine kraftvolle Arm- und Handgelenkstärkung, die ebenso fröhlich wie schön ist. Im vollen Ausdruck der Pose tritt Ihr unteres Bein auf den Boden, Ihr oberes Bein wird angehoben, bis es senkrecht zum Boden steht, und Ihr Oberkörper streckt sich.`,
+    level: ["intermediate", "pro"],
+    category: "Schultern",
+    benefit: "energizing",
+    next: ["Sphinx", "Child's Pose", "Downward-Facing Dog"],
+  },
+  {
+    id: 35,
+    sanskrit_name: "Salamba Bhujangasana",
+    english_name: "Sphinx",
+    img_url: "/images/sphinx.svg",
+    description: `Die Sphinx ist eine Variation der Cobra. Sie eignet sich besonders für Anfänger, die nocht nicht in eine so tiefe Rückbeuge einsteigen möchten. Gerade bei Schmerzen im unteren Rücken hat diese Haltung einen positiven Effekt. Die Wirbelsäule wird insgesamt gestärkt und der Brust- und Schulterbereich geöffnet. Die Durchblutung wird angeregt. Zu guter Letzt beruhigt die Sphinx Pose und wirkt Stress entgegen.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Rücken",
+    benefit: "calming",
+    next: ["Child's Pose", "Downward-Facing Dog"],
+  },
+  {
+    id: 36,
+    sanskrit_name: "Hanumanasana",
+    english_name: "Splits",
+    img_url: "/images/splits.svg",
+    description: `Der Spagat ist eine der fortgeschrittenen Yoga-Asanas, welche viel Zeit und Übung in Anspruch nimmt. Wie lange man braucht, um diese Haltung zu meistern ist von Person zu Person sehr unterschiedlich. Man sollte sie jedenfalls nicht erzwingen, denn die Verletzungsgefahr für ungeübte ist groß. Gerade die Oberschenkelrückseiten sind dabei gefährdet. Der Spagat ist also eine gute Übung, um neben körperlichen Fertigkeiten auch Geduld und Hingabe zu üben. `,
+    level: ["pro"],
+    category: "Beine",
+    benefit: "calming",
+    next: ["Seated Forward Bend", "King Pigeon"],
+  },
+
+  {
+    id: 37,
+    sanskrit_name: "Malasana",
+    english_name: "Squat",
+    img_url: "/images/squat.svg",
+    description: `Malasana - oder auch Yogi Squad genannt ist für viele Erwachsene eine echte Herausforderung. Die Hüft- und Beinmuskulatur bei vielen Menschen durch das häufige Sitzen stark verkürzt. Dies merken wir im Yogi Squad ganz besonders. Wer diese Asana regelmäßig praktiziert, kann aber eine spürbare Verbesserung erkennen. Die Haltung öffnet sanft die Hüften und bringt eine entspannte Rückendehnung mit sich. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Hüfte",
+    benefit: "calming",
+    next: ["Standing Forward Bend", "Downward-Facing Dog"],
+  },
+  {
+    id: 38,
+    sanskrit_name: "Uttanasana",
+    english_name: "Standing Forward Bend",
+    img_url: "/images/standingforwardfold.svg",
+    description: `In dieser Asana wird Ihre Wirbelsäule sowie die Oberschenkel-Rückseiten intensiv gedehnt. Diese intensive Vorbeuge eigent sich besonders bei Erschöpfungszuständen, Angstgefühlen oder Depressionen. Sie verlangsamt den Herzschlag und hat so eine beruhigende Wirkung auf das Nervensystem.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Rücken",
+    benefit: "calming",
+    next: ["Warrior One", "Plank", "Half-Moon"],
+  },
+  {
+    id: 39,
+    sanskrit_name: "Ashta Chandrasana",
+    english_name: "Crescent Moon",
+    img_url: "/images/standingstretch.svg",
+    description: `????`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Rücken",
+    benefit: "calming",
+    next: [""],
+  },
+  {
+    id: 40,
+    sanskrit_name: "Upavistha Konasana",
+    english_name: "Side Splits",
+    img_url: "/images/straddlesplit.svg",
+    description: `Lorem Ipsum`,
+    level: ["pro"],
+    category: "Hüfte",
+    benefit: "energizing",
+    next: [""],
+  },
+  {
+    id: 41,
+    sanskrit_name: "Vrksasana",
+    english_name: "Tree",
+    img_url: "/images/tree.svg",
+    description: `Lorem Ipsum`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Hüfte",
+    benefit: "calming",
+    next: ["Standing Forward Bend", ""],
+  },
+  {
+    id: 42,
+    sanskrit_name: "Trikonasana",
+    english_name: "Triangle",
+    img_url: "/images/triangle.svg",
+    description: `In dieser Asana bildet der Körper die Form eines langgestreckten Dreiecks. Der Rumpf und die Beine werden gedehnt. Die Haltung stärkt zudem die Bänder und verbessert ingesamt die Beweglichkeit. Die Auschrichtung der Schultern wird bei häufigem Üben verbessert. `,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Core",
+    benefit: "energizing",
+    next: [""],
+  },
+  {
+    id: 43,
+    sanskrit_name: "Urdhva Mukha Svsnssana",
+    english_name: "Upward-Facing Dog",
+    img_url: "/images/updog.svg",
+    description: `Lorem Ipsum`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Rücken",
+    benefit: "energizing",
+    next: ["Downward Facing Dog", "Plank", "Sphinx"],
+  },
+  {
+    id: 44,
+    sanskrit_name: "Virabhadrasana One",
+    english_name: "Warrior One",
+    img_url: "/images/warrior1.svg",
+    description: `Diese Übung ist der Haltung eines Kriegers nachempfunden. Die Asana stärkt die Wirbelsäule und trainiert Knie sowie Oberschenkel. Durch das Strecken der Arne wird der Brustkorb gedehnt. Die Rückenmuskulatur wird zudem gestärkt.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Beine",
+    benefit: "energizing",
+    next: ["Warrior Two", "Plank", "Reverse Warrior"],
+  },
+  {
+    id: 45,
+    sanskrit_name: "Virabhadrasana Two",
+    english_name: "Warrior Two",
+    img_url: "/images/warrior2.svg",
+    description: `Der Krieger zwei ist eine zentrale Haltung im Yoga und wird häufig in Verbindung mit dem Krieger eins praktiziert. Die Übung trainiert die Arme und Beine ebenso wie die Rumpfmuskulatur. Knie und Hüften werden beweglicher, wenn Sie diese Asana häufig praktizieren. Rückenschmerzen können durch diese Haltung spürbar gemindert werden.`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Beine",
+    benefit: "energizing",
+    next: ["Warrior Three", "Reverse Warrior"],
+  },
+  {
+    id: 46,
+    sanskrit_name: "Virabhadrasana Three",
+    english_name: "Warrior Three",
+    img_url: "/images/warrior3.svg",
+    description: `Lorem Ipsum`,
+    level: ["beginner", "intermediate", "pro"],
+    category: "Beine",
+    benefit: "energizing",
+    next: ["Pyramid", "Reverse Warrior", "Upward-Facing Dog"],
+  },
+  {
+    id: 47,
+    sanskrit_name: "Urdhva Dhanurasana",
+    english_name: "Wheel",
+    img_url: "/images/wheel.svg",
+    description: `Lorem Ipsum`,
+    level: ["intermediate", "pro"],
+    category: "Rücken",
+    benefit: "energizing",
+    next: ["Bridge"],
+  },
+  {
+    id: 48,
+    sanskrit_name: "Camatkarasana",
+    english_name: "Wild Thing",
+    img_url: "/images/wildthing.svg",
+    description: `Lorem Ipsum`,
+    level: ["intermediate", "pro"],
+    category: "Rücken",
+    benefit: "energizing",
+    next: ["Plank"],
+  },
+];
+
+export default asanas;
